requirements.txt
streamlit>=1.29.0
openai>=1.6.1
llama-cpp-python>=0.2.25

